no longer depend on (iso-transl)
(iso-transl) is a built-in package.  It seems that I'm no longer using
it.  I think that I used to use it when my keyboard layout has dead
accent keys.  But I no longer rely on dead accent keys that were not
practical and gave me issue sometimes in Emacs and sometimes in the
terminal.

* settings/key-bindings/kb.el:  no longer depend on (iso-transl).